+++ marker/forms/project.py
@@ -3,6 +3,7 @@ import re
 from sqlalchemy import select
 from wtforms import (
     DateTimeLocalField,
+    DecimalField,
     Form,
     HiddenField,
     SelectField,
@@ -21,6 +22,7 @@ from .select import (
     STAGES,
     STATUS,
     select_countries,
+    select_currencies,
     select_subdivisions,
 )
 from .ts import TranslationString as _
@@ -71,6 +73,27 @@ class ProjectForm(Form):
     delivery_method = SelectField(
         _("Project delivery method"), choices=PROJECT_DELIVERY_METHODS
     )
+    usable_area = DecimalField(
+        _("Usable area [m²]"),
+        validators=[Optional()],
+        places=2,
+    )
+    cubic_volume = DecimalField(
+        _("Cubic volume [m³]"),
+        validators=[Optional()],
+        places=2,
+    )
+    currency = SelectField(_("Currency"), choices=select_currencies())
+    value_net = DecimalField(
+        _("Net value"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_gross = DecimalField(
+        _("Gross value"),
+        validators=[Optional()],
+        places=2,
+    )
 
     def __init__(self, *args, request, **kwargs):
         super().__init__(*args, **kwargs)
@@ -108,15 +131,62 @@ class ProjectSearchForm(Form):
     )
     country = SelectField(_("Country"), choices=select_countries())
     website = StringField(_("Website"), filters=[strip_filter])
-    deadline = DateTimeLocalField(
-        _("Deadline"),
-        format="%Y-%m-%dT%H:%M",  # https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/datetime-local
+    deadline_from = DateTimeLocalField(
+        _("Deadline from"),
+        format="%Y-%m-%dT%H:%M",
+        validators=[Optional()],
+        widget=DateTimeLocalInput(),
+    )
+    deadline_to = DateTimeLocalField(
+        _("Deadline to"),
+        format="%Y-%m-%dT%H:%M",
         validators=[Optional()],
         widget=DateTimeLocalInput(),
     )
     stage = SelectField(_("Stage"), choices=STAGES)
     delivery_method = SelectField(
         _("Project delivery method"), choices=PROJECT_DELIVERY_METHODS
+    )
+    usable_area_from = DecimalField(
+        _("Usable area from [m²]"),
+        validators=[Optional()],
+        places=2,
+    )
+    usable_area_to = DecimalField(
+        _("Usable area to [m²]"),
+        validators=[Optional()],
+        places=2,
+    )
+    cubic_volume_from = DecimalField(
+        _("Cubic volume from [m³]"),
+        validators=[Optional()],
+        places=2,
+    )
+    cubic_volume_to = DecimalField(
+        _("Cubic volume to [m³]"),
+        validators=[Optional()],
+        places=2,
+    )
+    currency = SelectField(_("Currency"), choices=select_currencies())
+    value_net_from = DecimalField(
+        _("Net value from"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_net_to = DecimalField(
+        _("Net value to"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_gross_from = DecimalField(
+        _("Gross value from"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_gross_to = DecimalField(
+        _("Gross value to"),
+        validators=[Optional()],
+        places=2,
     )
     color = SelectField(_("Color"), choices=COLORS)
 
@@ -132,12 +202,65 @@ class ProjectFilterForm(Form):
     country = SelectField(_("Country"), choices=select_countries())
     website = HiddenField(_("Website"), filters=[strip_filter])
     deadline = HiddenField(_("Deadline"))
+    deadline_from = DateTimeLocalField(
+        _("Deadline from"),
+        format="%Y-%m-%dT%H:%M",
+        validators=[Optional()],
+        widget=DateTimeLocalInput(),
+    )
+    deadline_to = DateTimeLocalField(
+        _("Deadline to"),
+        format="%Y-%m-%dT%H:%M",
+        validators=[Optional()],
+        widget=DateTimeLocalInput(),
+    )
     stage = SelectField(_("Stage"), choices=STAGES)
     delivery_method = SelectField(
         _("Project delivery method"), choices=PROJECT_DELIVERY_METHODS
     )
     status = SelectField(_("Status"), choices=STATUS)
     color = SelectField(_("Color"), choices=COLORS)
+    currency = SelectField(_("Currency"), choices=select_currencies())
+    usable_area_from = DecimalField(
+        _("Usable area from [m²]"),
+        validators=[Optional()],
+        places=2,
+    )
+    usable_area_to = DecimalField(
+        _("Usable area to [m²]"),
+        validators=[Optional()],
+        places=2,
+    )
+    cubic_volume_from = DecimalField(
+        _("Cubic volume from [m³]"),
+        validators=[Optional()],
+        places=2,
+    )
+    cubic_volume_to = DecimalField(
+        _("Cubic volume to [m³]"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_net_from = DecimalField(
+        _("Net value from"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_net_to = DecimalField(
+        _("Net value to"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_gross_from = DecimalField(
+        _("Gross value from"),
+        validators=[Optional()],
+        places=2,
+    )
+    value_gross_to = DecimalField(
+        _("Gross value to"),
+        validators=[Optional()],
+        places=2,
+    )
     date_from = DateTimeLocalField(
         _("Date from"),
         format="%Y-%m-%dT%H:%M",